# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/customers/{id}:
    get:
      operationId: getAdminCustomer
      tags: [admin]
      summary: Get an admin customer profile and aggregate statistics
      security: [{ OperatorSession: [] }]
      parameters:
        - $ref: "#/components/parameters/AdminCustomerId"
      responses:
        "200":
          description: Customer detail
          content: { application/json: { schema: { $ref: "#/components/schemas/CustomerDetail" } } }
    patch:
      operationId: updateAdminCustomer
      tags: [admin]
      summary: Update app-owned customer profile and affiliate settings
      security: [{ OwnerSession: [] }]
      parameters:
        - $ref: "#/components/parameters/AdminCustomerId"
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/AdminCustomerUpdate" } } }
      responses:
        "200":
          description: Updated customer
          content: { application/json: { schema: { $ref: "#/components/schemas/CustomerDetail" } } }
  /admin/customers/{id}/referrals:
    get:
      operationId: getAdminCustomerReferrals
      tags: [admin]
      summary: List customer referrals
      security: [{ OperatorSession: [] }]
      parameters:
        - $ref: "#/components/parameters/AdminCustomerId"
        - name: page
          in: query
          schema: { type: number, minimum: 1, multipleOf: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: number, minimum: 1, maximum: 100, multipleOf: 1, default: 25 }
      responses:
        "200":
          description: Referrals
          content: { application/json: { schema: { $ref: "#/components/schemas/CustomerReferralPage" } } }
  /admin/customers/{id}/password-reset:
    post:
      operationId: resetAdminCustomerPassword
      tags: [admin]
      summary: Set a customer temporary password
      security: [{ OwnerSession: [] }]
      parameters: [{ $ref: "#/components/parameters/AdminCustomerId" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CustomerPasswordReset" } } }
      responses:
        "200":
          description: Action completed
          content: { application/json: { schema: { $ref: "#/components/schemas/CustomerActionResult" } } }
  /admin/customers/{id}/sessions-revoke:
    post:
      operationId: revokeAdminCustomerSessions
      tags: [admin]
      summary: Revoke all active customer sessions
      security: [{ OwnerSession: [] }]
      parameters: [{ $ref: "#/components/parameters/AdminCustomerId" }]
      responses:
        "200":
          description: All active sessions were revoked
          content: { application/json: { schema: { $ref: "#/components/schemas/CustomerActionResult" } } }
        "409":
          description: The customer has no linked Clerk identity or the Clerk client does not support session revocation
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "502":
          description: Clerk did not complete session revocation
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
  /admin/customers/{id}/email-confirm:
    post:
      operationId: confirmAdminCustomerEmail
      tags: [admin]
      summary: Initiate primary customer email verification when supported
      security: [{ OwnerSession: [] }]
      parameters: [{ $ref: "#/components/parameters/AdminCustomerId" }]
      responses:
        "200":
          description: Action completed
          content: { application/json: { schema: { $ref: "#/components/schemas/CustomerActionResult" } } }
  /admin/customers/{id}/suspend:
    post:
      operationId: suspendAdminCustomer
      tags: [admin]
      summary: Suspend a customer
      security: [{ OwnerSession: [] }]
      parameters: [{ $ref: "#/components/parameters/AdminCustomerId" }]
      responses:
        "200":
          description: Action completed
          content: { application/json: { schema: { $ref: "#/components/schemas/CustomerActionResult" } } }
  /admin/customers/{id}/activate:
    post:
      operationId: activateAdminCustomer
      tags: [admin]
      summary: Activate a customer
      security: [{ OwnerSession: [] }]
      parameters: [{ $ref: "#/components/parameters/AdminCustomerId" }]
      responses:
        "200":
          description: Action completed
          content: { application/json: { schema: { $ref: "#/components/schemas/CustomerActionResult" } } }
...
components:
  parameters:
...
    AdminCustomerId:
      name: id
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 255 }
...
  securitySchemes:
    CustomerSession:
      type: apiKey
      in: cookie
      name: __session
      description: Clerk customer session. Order ownership is derived from this authenticated session on the server.
    OperatorSession:
      type: apiKey
      in: cookie
      name: __session
      description: Clerk operator session. The authenticated user's verified email must match an active server-side operator account.
    OwnerSession:
      type: apiKey
      in: cookie
      name: __session
      description: Clerk owner session. The authenticated user must match the single active owner account in the server-side operator registry.
  schemas:
...
    ApiError:
      type: object
      required: [error, code, retryable, outcomeUnknown]
      properties:
        error: { type: string }
        code: { type: string }
        retryable: { type: boolean }
        outcomeUnknown: { type: boolean }
        orderId: { type: string }
...
    CustomerDetail:
      type: object
      required: [id, firstName, lastName, email, country, role, accountStatus, emailVerified, createdAt, lastActivity, affiliateCode, referralRate, totalOrders, doneOrders, balanceOwedUsd, referredUsers, totalReferralProfitUsd, sendVolume, receiveVolume, totalSalesUsd, refreshedAt]
      properties:
        id: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string, maxLength: 320 }
        country: { type: string }
        role: { type: string }
        accountStatus: { type: string, enum: [active, suspended] }
        emailVerified: { type: boolean }
        createdAt: { type: string, format: date-time }
        lastActivity: { type: string, format: date-time }
        affiliateCode: { type: ["string", "null"] }
        referralRate: { type: ["string", "null"], description: Per-affiliate decimal rate; null uses program default. }
        totalOrders: { type: number, minimum: 0, multipleOf: 1 }
        doneOrders: { type: number, minimum: 0, multipleOf: 1 }
        balanceOwedUsd: { type: string }
        referredUsers: { type: number, minimum: 0, multipleOf: 1 }
        totalReferralProfitUsd: { type: string }
        sendVolume:
          type: array
          items: { $ref: "#/components/schemas/CustomerAssetVolume" }
        receiveVolume:
          type: array
          items: { $ref: "#/components/schemas/CustomerAssetVolume" }
        totalSalesUsd:
          type: ["string", "null"]
          description: Sum of immutable affiliate USD valuations only; null when no authoritative valuation exists.
        refreshedAt: { type: string, format: date-time }
    CustomerAssetVolume:
      type: object
      required: [asset, amount]
      properties:
        asset: { type: string }
        amount: { type: string }
    CustomerReferral:
      type: object
      required: [id, name, country, ordersCount, registeredAt]
      properties:
        id: { type: string }
        name: { type: string }
        country: { type: string }
        ordersCount: { type: number, minimum: 0, multipleOf: 1 }
        registeredAt: { type: string, format: date-time }
    CustomerReferralPage:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/CustomerReferral" } }
        total: { type: number, minimum: 0, multipleOf: 1 }
        page: { type: number, minimum: 1, multipleOf: 1 }
        pageSize: { type: number, minimum: 1, maximum: 100, multipleOf: 1 }
    AdminCustomerUpdate:
      type: object
      minProperties: 1
      properties:
        firstName: { type: string, maxLength: 100 }
        lastName: { type: string, maxLength: 100 }
        country: { type: string, maxLength: 2 }
        role: { type: string, maxLength: 50 }
        referralCode: { type: string, pattern: '^(?:[A-HJ-NP-Z2-9]{8}|[A-HJ-NP-Z2-9]{12})$' }
        referralRate: { type: ["string", "null"], pattern: "^(0|0\\.[0-9]{1,18}|1|1\\.0{1,18})$" }
    CustomerPasswordReset:
      type: object
      required: [temporaryPassword]
      properties:
        temporaryPassword: { type: string, minLength: 12, maxLength: 128, writeOnly: true }
    CustomerActionResult:
      type: object
      required: [id, action, accountStatus, refreshedAt]
      properties:
        id: { type: string }
        action: { type: string }
        accountStatus: { type: string, enum: [active, suspended] }
        refreshedAt: { type: string, format: date-time }
...